debian/rules: build-arch depends on patch
Now that radiusd.conf is in freeradius-config, we need the
radiusd-to-freeradius.patch applied in build-arch.  If not, then
whether things are called freeradius or radiusd and what user is used
depends on whether the -B flag is passed to dpkg-buildpackage and on
whether the build is from git or from a tarball.  More consistency is
highly desirable.