trustrouter: permit trust-router-COI and error handling
Improve the error handling of the trust router integration.  Set
Error-Cause to indicate a proxy routing failure if we fail to get a
realm via trustrouter.  Include the error back from the trust router
in Reply-Message.

Set trust-router-coi in the trust router code rather than depending on
unlang to do so.  Permit a previously specified trust-router-coi in
the request to override the community in the configuration.  Unlang is
responsible for performing any authorization checks regarding
community.